refactor: move PropertySpec out of the feature-chainer package
property_spec.py is shared by PROPERTY_MAPPING (FeatureGroup) and
READER_OPTIONS (BaseInputData, which does not chain features), so
nesting it under feature_chainer/ misrepresented what it is. Relocate
it to components/property_spec.py as a sibling of feature_chainer/
and input_data/, and update every import across mloda/ and tests/.
