+++ tests/test_core/test_abstract_plugins/test_components/feature_chainer/test_contained_raise_log_no_exception_object.py
@@ -16,7 +16,7 @@ from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser
     FeatureChainParser,
     PropertyValueRejection,
 )
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec
 from mloda.user import Options
 
 PARSER_LOGGER_NAME = FeatureChainParser.__module__
