+++ tests/test_core/test_abstract_plugins/test_components/feature_chainer/test_property_spec_author_time.py
@@ -13,7 +13,7 @@ from __future__ import annotations
 
 import pytest
 
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec, property_spec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec, property_spec
 
 
 class TestAuthorTimeTypeErrors:
