+++ tests/test_core/test_abstract_plugins/test_components/feature_chainer/test_property_spec_hard_break.py
@@ -34,7 +34,7 @@ from mloda.core.abstract_plugins.components import default_options_key as defaul
 from mloda.core.abstract_plugins.components.default_options_key import DefaultOptionKeys
 from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser import FeatureChainParser
 from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser_mixin import FeatureChainParserMixin
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec, property_spec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec, property_spec
 from mloda.core.abstract_plugins.components.feature_set import FeatureSet
 from mloda.core.abstract_plugins.components.options import Options
 from mloda.core.abstract_plugins.feature_group import FeatureGroup
