+++ tests/test_core/test_abstract_plugins/test_components/feature_chainer/test_retire_transitional_seams.py
@@ -30,7 +30,7 @@ from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser
     option_key_is_present,
 )
 from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser_mixin import FeatureChainParserMixin
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec
 from mloda.core.abstract_plugins.components.options import Options
 
 _FLAGLESS_SPEC = PropertySpec("a key with the default presence rule")
