+++ tests/test_core/test_abstract_plugins/test_components/test_declaration_surface.py
@@ -10,7 +10,7 @@ from typing import Any, ClassVar
 
 import pytest
 
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec
 from mloda.core.abstract_plugins.components.input_data.base_input_data import BaseInputData
 from mloda.core.abstract_plugins.components.declaration_surface import (
     DeclarationSurface,
