+++ tests/test_core/test_abstract_plugins/test_components/test_reader_option_contained_raise_log_no_exception_object.py
@@ -11,7 +11,7 @@ from typing import Any
 
 import pytest
 
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec
 from mloda.core.abstract_plugins.components.input_data.base_input_data import BaseInputData
 from mloda.user import Options
