+++ tests/test_core/test_prepare/test_match_abort_escalation.py
@@ -18,7 +18,7 @@ from typing import ClassVar, Optional, TypeVar
 from mloda.core.abstract_plugins.components.data_access_collection import DataAccessCollection
 from mloda.core.abstract_plugins.components.feature import Feature
 from mloda.core.abstract_plugins.components.feature_chainer.feature_chain_parser_mixin import FeatureChainParserMixin
-from mloda.core.abstract_plugins.components.feature_chainer.property_spec import PropertySpec, property_spec
+from mloda.core.abstract_plugins.components.property_spec import PropertySpec, property_spec
 from mloda.core.abstract_plugins.components.feature_name import FeatureName
 from mloda.core.abstract_plugins.components.options import Options
 from mloda.core.abstract_plugins.components.utils import escalate_match_abort, is_match_abort
